add docs and merge JdbcDataFetcher and JdbcDataPreFetcher
