+++ extensions-core/lookups-cached-single/pom.xml
@@ -61,6 +61,11 @@
       <artifactId>stringtemplate</artifactId>
       <version>3.2</version>
     </dependency>
+    <dependency>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-collections4</artifactId>
+      <version>4.0</version>
+    </dependency>
     <!-- Tests -->
     <dependency>
       <groupId>junit</groupId>
