+++ extensions-core/lookups-cached-single/src/main/java/io/druid/server/lookup/jdbc/JdbcDataFetcher.java
@@ -24,11 +24,14 @@ package io.druid.server.lookup.jdbc;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
+import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
-
+import com.google.common.collect.Maps;
+import com.metamx.common.Pair;
 import io.druid.java.util.common.logger.Logger;
 import io.druid.metadata.MetadataStorageConnectorConfig;
-import io.druid.server.lookup.DataFetcher;
+import io.druid.server.lookup.PrefetchableFetcher;
+import org.apache.commons.collections4.CollectionUtils;
 import org.skife.jdbi.v2.DBI;
 import org.skife.jdbi.v2.Handle;
 import org.skife.jdbi.v2.TransactionCallback;
@@ -40,8 +43,10 @@ import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.List;
 import java.util.Map;
+import java.util.NavigableSet;
+import java.util.TreeSet;
 
-public class JdbcDataFetcher implements DataFetcher<String, String>
+public class JdbcDataFetcher extends PrefetchableFetcher<String, String>
 {
   private static final Logger LOGGER = new Logger(JdbcDataFetcher.class);
   private static final int DEFAULT_STREAMING_FETCH_SIZE = 1000;
@@ -56,18 +61,25 @@ public class JdbcDataFetcher implements DataFetcher<String, String>
   private final String valueColumn;
   @JsonProperty
   private final int streamingFetchSize;
+  @JsonProperty
+  private final List<String> prefetchRanges;
 
   private final String fetchAllQuery;
   private final String fetchQuery;
+  private final String prefetchQueryFromTo;
+  private final String prefetchQueryFromOnly;
+  private final String prefetchQueryToOnly;
   private final String reverseFetchQuery;
   private final DBI dbi;
+  private final NavigableSet<String> sortedRanges;
 
   public JdbcDataFetcher(
       @JsonProperty("connectorConfig") MetadataStorageConnectorConfig connectorConfig,
       @JsonProperty("table") String table,
       @JsonProperty("keyColumn") String keyColumn,
       @JsonProperty("valueColumn") String valueColumn,
-      @JsonProperty("streamingFetchSize") Integer streamingFetchSize
+      @JsonProperty("streamingFetchSize") Integer streamingFetchSize,
+      @JsonProperty("prefetchRanges") List<String> rangeKeys
   )
   {
     this.connectorConfig = Preconditions.checkNotNull(connectorConfig, "connectorConfig");
@@ -76,6 +88,8 @@ public class JdbcDataFetcher implements DataFetcher<String, String>
     this.table = Preconditions.checkNotNull(table, "table");
     this.keyColumn = Preconditions.checkNotNull(keyColumn, "keyColumn");
     this.valueColumn = Preconditions.checkNotNull(valueColumn, "valueColumn");
+    this.prefetchRanges = rangeKeys;
+    this.sortedRanges = new TreeSet<>(this.prefetchRanges);
 
     this.fetchAllQuery = String.format(
         "SELECT %s, %s FROM %s",
@@ -85,6 +99,28 @@ public class JdbcDataFetcher implements DataFetcher<String, String>
     );
     this.fetchQuery = String.format(
         "SELECT %s FROM %s WHERE %s = :val",
+        this.valueColumn,
+        this.table,
+        this.keyColumn
+    );
+    this.prefetchQueryFromTo = String.format(
+        "SELECT %s, %s FROM %s WHERE '%%s' <= %s AND %s < '%%s'",
+        this.keyColumn,
+        this.valueColumn,
+        this.table,
+        this.keyColumn,
+        this.keyColumn
+    );
+    this.prefetchQueryFromOnly = String.format(
+        "SELECT %s, %s FROM %s WHERE %s < '%%s'",
+        this.keyColumn,
+        this.valueColumn,
+        this.table,
+        this.keyColumn
+    );
+    this.prefetchQueryToOnly = String.format(
+        "SELECT %s, %s FROM %s WHERE '%%s' <= %s",
+        this.keyColumn,
         this.valueColumn,
         this.table,
         this.keyColumn
@@ -115,9 +151,9 @@ public class JdbcDataFetcher implements DataFetcher<String, String>
                                    ) throws Exception
                                    {
                                      return handle.createQuery(fetchAllQuery)
-                                                  .setFetchSize(streamingFetchSize)
-                                                  .map(new KeyValueResultSetMapper(keyColumn, valueColumn))
-                                                  .list();
+                                         .setFetchSize(streamingFetchSize)
+                                         .map(new KeyValueResultSetMapper(keyColumn, valueColumn))
+                                         .list();
                                    }
 
                                  }
@@ -134,9 +170,9 @@ public class JdbcDataFetcher implements DataFetcher<String, String>
           public List<String> inTransaction(Handle handle, TransactionStatus status) throws Exception
           {
             return handle.createQuery(fetchQuery)
-                         .bind("val", key)
-                         .map(StringMapper.FIRST)
-                         .list();
+                .bind("val", key)
+                .map(StringMapper.FIRST)
+                .list();
           }
         }
     );
@@ -162,12 +198,44 @@ public class JdbcDataFetcher implements DataFetcher<String, String>
       public List<String> inTransaction(Handle handle, TransactionStatus status) throws Exception
       {
         return handle.createQuery(reverseFetchQuery)
-                     .bind("val", value)
-                     .map(StringMapper.FIRST)
-                     .list();
+            .bind("val", value)
+            .map(StringMapper.FIRST)
+            .list();
       }
     });
     return results;
+  }
+
+  @Override
+  public Map<String, String> prefetch(String key)
+  {
+    if (prefetchRanges == null) {
+      return ImmutableMap.of();
+    }
+
+    final Pair<String, String> range = getRange(key);
+    final String queryString =
+        (range.lhs == null) ? String.format(prefetchQueryToOnly, range.rhs) :
+            (range.rhs == null) ? String.format(prefetchQueryFromOnly, range.lhs)
+                : String.format(prefetchQueryFromTo, range.lhs, range.rhs);
+
+    return inReadOnlyTransaction(
+        new TransactionCallback<Map<String, String>>()
+        {
+          @Override
+          public Map<String, String> inTransaction(Handle handle, TransactionStatus status) throws Exception
+          {
+            List<Map<String, Object>> rowList = handle.createQuery(queryString)
+                .setFetchSize(streamingFetchSize)
+                .list();
+            Map<String, String> rowMap = Maps.newHashMap();
+            for (Map<String, Object> row: rowList) {
+              rowMap.put((String) row.get(keyColumn), (String) row.get(valueColumn));
+            }
+            return rowMap;
+          }
+        }
+    );
   }
 
   @Override
@@ -191,8 +259,21 @@ public class JdbcDataFetcher implements DataFetcher<String, String>
     if (!keyColumn.equals(that.keyColumn)) {
       return false;
     }
+    if (!CollectionUtils.isEqualCollection(prefetchRanges, that.prefetchRanges)) {
+      return false;
+    }
     return valueColumn.equals(that.valueColumn);
 
+  }
+
+  private Pair<String, String> getRange(String key)
+  {
+    String start = sortedRanges.floor(key);
+    String end = sortedRanges.higher(key);
+
+    Preconditions.checkArgument(start != null || end != null, "No matching range");
+
+    return new Pair<String, String>(start, end);
   }
 
   private DBI getDbi()
@@ -227,5 +308,4 @@ public class JdbcDataFetcher implements DataFetcher<String, String>
         }
     );
   }
-
 }
